Refactored services > helpers
Renamed package services to helpers
