+++ tkinter_mvc_app/helpers/const.py
